Moved some common modules into "common"
This is just to make development a bit faster. One can symlink pdreq for the time being to make API calls a lot easier.

In addition to this, I've added a new feature to apireq: better errors, and ability to loop through pagination and get all objects of a given type.